alterações

+++ Lista_1/questao_1.java
@@ -1,8 +1,7 @@
 public class questao_1{
     public static void main(String[] args){
         float B = 7.5f; float H = 12.3f;
-       // Scanner scanner = new Scanner(system.in);
-        float area = 0.0f; float perimetro = 0.0f;
+        double area = 0.0d; double perimetro = 0.0d;
         area = B * H; 
         perimetro = (B * 2) + (H * 2);
         System.out.println("A área é: " + area + " cm");
